backend/requirements.txt
fastapi
uvicorn
anthropic
openai
httpx
openpyxl
pdfplumber
python-docx
python-multipart
pytest
python-dotenv
